requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
databricks-cli==0.11.0
databricks-connect==6.5.1
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.1
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-JSON==0.3.4
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.3
Flask-User==1.0.2.2
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
passlib==1.7.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
py4j==0.10.7
pycparser==2.20
python-dotenv==0.14.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
tabulate==0.8.7
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
xmltodict==0.12.0


beautifulsoup4
pandas
sklearn